Create login & logout backend for user

+++ src/utils/cloudinary.js
@@ -1,24 +1,27 @@
 import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
-
+import "dotenv/config";
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key:process.env.CLOUDINARY_API_KEY,
     api_secret: process.env.CLOUDINARY_API_SECRET
 });
-
+console.log(cloudinary.config());
 const uploadOnCloudinary = async (localFilePath) => {
     try{
 if(!localFilePath) return null
-cloudinary.uploader.upload(localFilePath, {
+const response = await cloudinary.uploader.upload(localFilePath, {
     resource_type: "auto"
 })
-   console.log("file is uploaded on cloudinary",
-    response.url
-   );
+   //console.log("file is uploaded on cloudinary",response.url);
+   fs.unlinkSync(localFilePath)
    return response;
-    } catch (error){
-        fs.unlinkSync(localFilePath)
-        return null;
+    } catch (error) {
+    console.log("Cloudinary Error:", error);
+    if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
+    }
+    return null;
 }
 }
+export {uploadOnCloudinary}
